+++ ranking/src/pagerank.cpp
@@ -1,5 +1,6 @@
 #include "pagerank.h"
 
-PageRank::PageRank(float *matrix, int rows, int columns) : matrix_(matrix), num_rows_(rows), num_columns_(columns) {
+PageRank::PageRank(float* matrix, int rows, int columns)
+    : matrix_(matrix), num_rows_(rows), num_columns_(columns) {
 
-};
+      };